Rename DataSet.[c|h]pp to DataSetHDF5.[c|h]pp
The HDF5 DataSet, not the toplevel one.
