+++ src/hdf5/DataArrayHDF5.cpp
@@ -9,7 +9,7 @@
 #include <nix/util/util.hpp>
 
 #include <nix/hdf5/DataArrayHDF5.hpp>
-#include <nix/hdf5/DataSet.hpp>
+#include <nix/hdf5/DataSetHDF5.hpp>
 #include <nix/hdf5/DimensionHDF5.hpp>
 
 using namespace std;
